Rename module and sort references map for deterministic ordering
